Move `OwningShadowNodeFragment` to a separate file
