+++ cpp/react/renderer/components/RNLiveMarkdownSpec/MarkdownTextInputDecoratorShadowNode.cpp
@@ -1,8 +1,5 @@
 #if defined(RCT_NEW_ARCH_ENABLED) || defined(ANDROID)
 
-#include <react/renderer/core/LayoutContext.h>
-
-#include "MarkdownShadowFamilyRegistry.h"
 #include "MarkdownTextInputDecoratorShadowNode.h"
 
 using namespace expensify::livemarkdown;
