+++ cpp/react/renderer/components/RNLiveMarkdownSpec/MarkdownTextInputDecoratorShadowNode.h
@@ -3,7 +3,7 @@
 
 #include "MarkdownShadowFamilyRegistry.h"
 #include "MarkdownTextInputDecoratorState.h"
-#include <jsi/jsi.h>
+#include "OwningShadowNodeFragment.h"
 #include <react/renderer/components/RNLiveMarkdownSpec/EventEmitters.h>
 #include <react/renderer/components/RNLiveMarkdownSpec/Props.h>
 #include <react/renderer/components/view/ConcreteViewShadowNode.h>
@@ -12,20 +12,6 @@ using namespace expensify::livemarkdown;
 
 namespace facebook {
 namespace react {
-
-struct OwningShadowNodeFragment {
-  Props::Shared props;
-  ShadowNode::SharedListOfShared children;
-  State::Shared state;
-  
-  operator ShadowNodeFragment() const {
-    return ShadowNodeFragment {
-      .props = props,
-      .children = children,
-      .state = state
-    };
-  }
-};
 
 JSI_EXPORT extern const char MarkdownTextInputDecoratorViewComponentName[];
